backend/Building AI Summarization and Unified Calendar Backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.12.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
python-docx>=1.1.0
supabase>=2.3.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
requests>=2.31.0
